[#17] Refactor: Daily challenge 관련 entity 수정
- DailyChallenge entity 수정
- DailyAuth entity 수정
- DailyParticipate entity 수정
- ChallengeStatus enum 추가
