+++ src/main/java/com/itoxi/petnuri/domain/dailychallenge/entity/DailyChallenge.java
@@ -1,5 +1,6 @@
-package com.itoxi.petnuri.domain.dailychallenge;
+package com.itoxi.petnuri.domain.dailychallenge.entity;
 
+import com.itoxi.petnuri.domain.dailychallenge.type.ChallengeStatus;
 import com.itoxi.petnuri.global.common.BaseTimeEntity;
 import lombok.*;
 
@@ -10,7 +11,7 @@ import java.util.List;
 /**
  * author         : matrix
  * date           : 2023-09-13
- * description    :
+ * description    : 데일리 챌린지
  */
 @Entity
 @Getter
@@ -32,10 +33,15 @@ public class DailyChallenge extends BaseTimeEntity {
     @Column(nullable = false)
     private Long payment;       // 챌린지 인증 완료시 지급 포인트
 
-    @Temporal(TemporalType.TIMESTAMP)
+    @Column(nullable = false)
     private LocalDateTime startDate;    // 챌린지 시작 일자 : 2023-09-12 00:00:00
-    @Temporal(TemporalType.TIMESTAMP)
-    private LocalDateTime endDate;      // 챌린지 종료 일자 : 2023-09-12 23:59:59
+
+    //Todo: 데일리 챌린지에 종료 일자가 필요한지 생각해 보기
+    @Column(nullable = false)
+    private LocalDateTime endDate;      // 챌린지 종료 일자 : 9999-12-31 23:59:59
+
+    @Enumerated(value = EnumType.STRING)
+    private ChallengeStatus challengeStatus;
 
     @OneToMany(mappedBy = "dailyChallenge")
     private List<DailyParticipate> dailyParticipate = new ArrayList<>(); // 챌린지 참여 매핑
