+++ src/main/java/com/itoxi/petnuri/domain/dailychallenge/entity/DailyParticipate.java
@@ -1,16 +1,13 @@
-package com.itoxi.petnuri.domain.dailychallenge;
+package com.itoxi.petnuri.domain.dailychallenge.entity;
 
-import com.itoxi.petnuri.domain.member.entity.Member;
 import com.itoxi.petnuri.global.common.BaseTimeEntity;
 import lombok.*;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 /**
  * author         : matrix
  * date           : 2023-09-13
- * description    :
+ * description    : 데일리 챌린지 참여
  */
 @Entity
 @Getter
@@ -30,19 +27,9 @@ public class DailyParticipate extends BaseTimeEntity {
     private DailyChallenge dailyChallenge;  // 챌린지 참여
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Transient
-    private Boolean autoSave;  // 펫톡 게시판 자동 등록 여부
+    @JoinColumn(name = "daily_auth_id")
+    private DailyAuth dailyAuth;
 
     @Transient
     private Boolean status;     // 인증글 작성 여부
-
-    @OneToMany(mappedBy = "dailyParticipate")
-    private List<DailyAuth> dailyAuthList = new ArrayList<>();
-
 }
