+++ src/main/java/com/itoxi/petnuri/domain/dailychallenge/type/ChallengeStatus.java
@@ -1,0 +1,10 @@
+package com.itoxi.petnuri.domain.dailychallenge.type;
+
+/**
+ * author         : matrix
+ * date           : 2023-09-16
+ * description    :
+ */
+public enum ChallengeStatus {
+    READY, OPENED, CLOSED
+}
